Arduino XBee: Mid-demo nodes
New Arduino project for each the bike, corner and master node using
series 1 XBees in API mode. Bike sends a broadcast, two corners
receive and send the RSSI from the broadcast to the master, along with
the nodeID of the bike node. Plenty to clean up, but the master seems to
be deciding the binary location of the bike fairly well based on the
RSSI from the two corners.