Bike Mole: Added IO response on reception
When the bike mole receives a packet from the master, it now reads its
relative location and sets the appropriate LEDs for the vehicles
location. A timer is started at the expiry of which the LEDs are turned
off. Buzzer code is commented out as the buzzer is not yet attached to
the hardware.